\name{ColesData}
\docType{data}
\alias{ColesData}
\title{Annual Maximum Sea Levels at Port Pirie, South Australia}
\description{
This dataset is taken from Coles (2001) (also see references therein), and consists of 64 sea level (in meters) yearly maxima for the time period 1923-1987.
}
\usage{data(ColesData)}
\format{A \code{\link{data.frame}} containing two columns named \code{year} and \code{sea.level} (in meters).}
\source{Coles (2001), page 4 (also see references therein).}
\references{
Coles, S. (2001). An introduction to statistical modeling of extreme values. Springer.
}
\keyword{datasets}
