\name{FAmle-package}
\Rdversion{1.2}
\alias{FAmle-package}
\alias{FAmle}
\docType{package}
\title{Maximum Likelihood and Bayesian Estimation of Univariate Probability Distributions}
\description{This package contains a series of functions that might be useful in carrying out maximum likelihood and Bayesian estimations of univariate probability distributions.}

\author{
Francois Aucoin (author and original maintainer),
Thomas Petzoldt (actual maintainer, applied formal changes to pass CRANs
package check). Many thanks to the original author for his agreement.
}

\keyword{package}

