\name{station01AJ010}
\docType{data}
\alias{station01AJ010}
\title{Annual Maximum Daily Mean Flow Data (NB, Canada)}
\description{
This dataset is taken from the HYDAT database, and corresponds to realized values of annual maximum daily mean flows (in \eqn{m^3/s}).
}
\usage{data(station01AJ010)}
\format{A vector of observations.}
\source{Hydrometric station 01AJ010}
\references{
Environment Canada: \url{https://wateroffice.ec.gc.ca/}
}
\keyword{datasets}
