\name{yarns}
\docType{data}
\alias{yarns}
\title{Yarns Failure Data}
\description{
This dataset is taken from Gamerman and Lopes (2006) (also see references therein), and consists of 100 cycles-to-failure times for airplane yarns.
}
\usage{data(yarns)}
\format{A \code{\link{data.frame}} object - one column of 100 observations.}
\source{Gamerman and Lopes (2006), page 255 (also see references therein).}
\references{
Gamerman, D., and Lopes H.F. (2006). Markov Chain Monte Carlo: Stochastic simulation for Bayesian inference. 2nd edition, Chapman & Hall/CRC.
}
\keyword{datasets}
