\name{FAmle-package}
\Rdversion{1.2}
\alias{FAmle-package}
\alias{FAmle}
\docType{package}
\title{Maximum Likelihood and Bayesian Estimation of Univariate Probability Distributions}
\description{This package contains a series of functions that might be useful in carrying out maximum likelihood and Bayesian estimations of univariate probability distributions.}

\author{
Francois Aucoin

Maintainer: Francois Aucoin <frank.aucoin@gmail.com>

Site: \url{https://sites.google.com/site/frunkaucoin/}
}

\keyword{package}

